refactor(shared): single home for yjs token scheme + worker log facade
The backend's token signer and the relay's verifier implemented the same
HMAC scheme in two files with the payload typed twice; both are now thin
wrappers over shared/utils/yjs-token. The byte-identical worker pino
setups collapse into createWorkerLog(serviceSuffix, env).
