+++ shared/package.json
@@ -37,7 +37,8 @@
     "picocolors": "^1.1.1",
     "pino": "^10.3.1",
     "pino-opentelemetry-transport": "^4.0.2",
-    "uuidv7": "^1.2.1"
+    "uuidv7": "^1.2.1",
+    "zod": "^4.4.3"
   },
   "devDependencies": {
     "@typescript/native-preview": "7.0.0-dev.20260707.2",
